+++ packages/ui/src/index.ts
@@ -10,3 +10,5 @@ export {
   type CardBodyProps,
   type CardFooterProps,
 } from './Card'
+export { OTPInput, type OTPInputProps } from './OTPInput'
+export { ResendTimer, type ResendTimerProps } from './ResendTimer'